using size_t

+++ code/src/newneunet/learning.c
@@ -98,16 +98,16 @@ void backpropagate(neuralnet_t* net, const float* input, const float* target_out
 	float** errors = SAFE_MALLOC(net->layer_count * sizeof(float*));
 
 	// Calculating direct errors of last layer
-	uint32_t last = net->layer_count - 1;
+	size_t last = net->layer_count - 1;
 	
 	errors[last] = SAFE_MALLOC(net->neurons_per_layer[last] * sizeof(float));
 
-	for(uint32_t to = 0; to < net->neurons_per_layer[last]; ++to){
+	for(size_t to = 0; to < net->neurons_per_layer[last]; ++to){
 
 		errors[last][to] = calculate_sigmoid_error(touts[to], fout->values[last][to]);
 
 		// Correcting edge weights of gap right before the last layer of neurons
-		for(uint32_t from = 0; from < net->neurons_per_layer[last - 1]; ++from){
+		for(size_t from = 0; from < net->neurons_per_layer[last - 1]; ++from){
 
 			net->edges[last - 1][from][to] += learn_rate * errors[last][to] * fout->values[last - 1][from];
 
@@ -116,15 +116,15 @@ void backpropagate(neuralnet_t* net, const float* input, const float* target_out
 	}
 
 	// Calculating intermediate errors of all remaining layers
-	for(uint32_t layer = net->layer_count - 2; layer > 0; --layer){
+	for(size_t layer = net->layer_count - 2; layer > 0; --layer){
 
 		errors[layer] = SAFE_MALLOC(net->neurons_per_layer[layer] * sizeof(float));
 
-		for(uint32_t current = 0; current < net->neurons_per_layer[layer]; ++current){
+		for(size_t current = 0; current < net->neurons_per_layer[layer]; ++current){
 
 			errors[layer][current] = 0.0f;
 
-			for(uint32_t right = 0; right < net->neurons_per_layer[layer + 1]; ++right){
+			for(size_t right = 0; right < net->neurons_per_layer[layer + 1]; ++right){
 
 				errors[layer][current] += learn_rate * errors[layer + 1][right] * net->edges[layer][current][right];
 
@@ -133,7 +133,7 @@ void backpropagate(neuralnet_t* net, const float* input, const float* target_out
 			errors[layer][current] *= calculate_sigmoid_error_term(fout->values[layer][current]);
 
 			// Correcting edge weights of gap right before the present layer of neurons
-			for(uint32_t left = 0; left < net->neurons_per_layer[layer - 1]; ++left){
+			for(size_t left = 0; left < net->neurons_per_layer[layer - 1]; ++left){
 
 				net->edges[layer - 1][left][current] += errors[layer][current] * fout->values[layer - 1][left];
 
@@ -147,7 +147,7 @@ void backpropagate(neuralnet_t* net, const float* input, const float* target_out
 	free_full_output(fout);
 	free(ins);
 	free(touts);
-	for(uint32_t layer = 1; layer < net->layer_count; ++layer)
+	for(size_t layer = 1; layer < net->layer_count; ++layer)
 		free(errors[layer]);
 	free(errors);
